refactor: separating the build logic of compose
